refactor: change add folders to improve the structuring of the project
