+++ .eslintrc.json
@@ -15,6 +15,7 @@
       "@typescript-eslint/explicit-function-return-type": "off",
       "@typescript-eslint/no-misused-promises": "off",
       "@typescript-eslint/strict-boolean-expressions": "off",
-      "@typescript-eslint/prefer-nullish-coalescing": "off"
+      "@typescript-eslint/prefer-nullish-coalescing": "off",
+      "no-useless-escape": "off"
     }
 }
