+++ package.json
@@ -11,6 +11,7 @@
     "lint:fix": "yarn run lint --fix"
   },
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
     "@types/express": "^4.17.14",
     "@types/mongoose": "^5.11.97",
     "@types/multer": "^1.4.7",
@@ -29,6 +30,7 @@
     "typescript": "*"
   },
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "express": "^4.18.2",
     "mongoose": "^6.7.2",
     "multer": "^1.4.5-lts.1"
