+++ src/app/router.ts
@@ -3,13 +3,10 @@ import { Router } from 'express'
 import { categoryRouter } from './routes/categoriesrouter'
 import { productRouter } from './routes/productsrouter'
 import { orderRouter } from './routes/ordersrouter'
-
-import { cancelOrder } from './usecases/orders/cancelOrder'
+import { userRouter } from './routes/usersrouter'
 
 export const router = Router()
 router.use(categoryRouter)
 router.use(productRouter)
 router.use(orderRouter)
-
-// Delete / cancelar o pedido
-router.delete('/orders/:orderId', cancelOrder)
+router.use(userRouter)
