+++ src/index.ts
@@ -6,7 +6,7 @@ import { router } from './app/router'
 mongoose.connect('mongodb://localhost:27017')
   .then(() => {
     const app = express()
-    app.use('/uploads', express.static(path.resolve(__dirname, '..', '..', 'uploads')))
+    app.use('/uploads', express.static(path.resolve(__dirname, '..', 'uploads')))
     app.use(express.json())
     app.use(router)
